Setting to customize Metadata URLs in sitemap
Add a new settinng `system/server/sitemapLinkUrl` to customize the HTML links to the
metadata documents in sitemap service. Customized URLs must include the pattern {{UUID}} that
will be replaced by the actual metadata UUID. If the setting is empty or not present GN default
value for the URL will be used.